+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/circuit/operator/DBSPSourceBaseOperator.java
@@ -33,19 +33,17 @@ import javax.annotation.Nullable;
 public abstract class DBSPSourceBaseOperator extends DBSPOperator {
     public final String tableName;
 
-    /**
-     * Create a DBSP operator that is a source to the dataflow graph.
-     * @param node        Calcite node for the statement creating the table
-     *                    that this node is created from.
-     * @param isMultiset  True if the source data can be a multiset.
-     * @param outputType  Type of table.
-     * @param comment     A comment describing the operator.
-     * @param tableName   The name of the table that this operator is created from.
-     */
-    public DBSPSourceBaseOperator(
-            CalciteObject node,
-            DBSPType outputType, boolean isMultiset, @Nullable String comment,
-            String tableName) {
+    /** Create a DBSP operator that is a source to the dataflow graph.
+     *
+     * @param node       Calcite node for the statement creating the table
+     *                   that this node is created from.
+     * @param outputType Type of table.
+     * @param isMultiset True if the source data can be a multiset.
+     * @param tableName  The name of the table that this operator is created from.
+     * @param comment    A comment describing the operator. */
+    protected DBSPSourceBaseOperator(
+            CalciteObject node, DBSPType outputType, boolean isMultiset,
+            String tableName, @Nullable String comment) {
         super(node, "source " + tableName, null, outputType, isMultiset, comment);
         this.tableName = tableName;
     }