+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/circuit/operator/DBSPSourceTableOperator.java
@@ -1,6 +1,6 @@
 package org.dbsp.sqlCompiler.circuit.operator;
 
-import org.dbsp.sqlCompiler.compiler.InputTableMetadata;
+import org.dbsp.sqlCompiler.compiler.TableMetadata;
 import org.dbsp.sqlCompiler.compiler.errors.SourcePositionRange;
 import org.dbsp.sqlCompiler.compiler.frontend.calciteObject.CalciteObject;
 import org.dbsp.sqlCompiler.ir.type.DBSPType;
@@ -15,23 +15,24 @@ public abstract class DBSPSourceTableOperator extends DBSPSourceBaseOperator {
     public final CalciteObject sourceName;
     // Note: the metadata is not transformed after being set.
     // In particular, types are not rewritten.
-    public final InputTableMetadata metadata;
+    public final TableMetadata metadata;
 
     /**
      * Create a DBSP operator that is a source to the dataflow graph.
-     * @param node        Calcite node for the statement creating the table
-     *                    that this node is created from.
-     * @param sourceName  Calcite node for the identifier naming the table.
-     * @param outputType  Type of table.
-     * @param isMultiset  True if the source can produce multiset values.
-     * @param comment     A comment describing the operator.
-     * @param name        The name of the table that this operator is created from.
+     *
+     * @param node       Calcite node for the statement creating the table
+     *                   that this node is created from.
+     * @param sourceName Calcite node for the identifier naming the table.
+     * @param outputType Type of table.
+     * @param isMultiset True if the source can produce multiset values.
+     * @param name       The name of the table that this operator is created from.
+     * @param comment    A comment describing the operator.
      */
-    public DBSPSourceTableOperator(
+    protected DBSPSourceTableOperator(
             CalciteObject node, CalciteObject sourceName,
             DBSPType outputType, DBSPTypeStruct originalRowType, boolean isMultiset,
-            @Nullable String comment, InputTableMetadata metadata, String name) {
-        super(node, outputType, isMultiset, comment, name);
+            TableMetadata metadata, String name, @Nullable String comment) {
+        super(node, outputType, isMultiset, name, comment);
         this.originalRowType = originalRowType;
         this.sourceName = sourceName;
         this.metadata = metadata;